feature/configuration/ 아티팩트 오타 수정 bookkurin -> bokkurin
